requirements.txt
undetected-chromedriver
python-dotenv
selenium